requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-babel==4.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
pytz==2025.2
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
email_validator==2.1.1